requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
pypdf>=3.17.0
sentence-transformers>=2.2.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0